requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
protobuf==4.21.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.1.2
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.10
Werkzeug==2.1.2
